# client/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ChatMessageMetadata": {
        "description": "UIMessage `metadata` of an assistant turn that never completed:\n`interrupted` alone for a stop/disconnect, plus `errorText` when the run\nfailed.",
        "properties": {
          "errorText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errortext"
          },
          "interrupted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupted"
          }
        },
        "title": "ChatMessageMetadata",
        "type": "object"
      },
      "ChatModelOption": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/LLMProvider"
          },
          "supports_reasoning_effort": {
            "title": "Supports Reasoning Effort",
            "type": "boolean"
          },
          "supports_vision": {
            "title": "Supports Vision",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "provider",
          "supports_reasoning_effort",
          "supports_vision"
        ],
        "title": "ChatModelOption",
        "type": "object"
      },
      "ChatModelsResponse": {
        "properties": {
          "default": {
            "title": "Default",
            "type": "string"
          },
          "default_provider": {
            "$ref": "#/components/schemas/LLMProvider"
          },
          "default_reasoning_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Reasoning Effort"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ChatModelOption"
            },
            "title": "Models",
            "type": "array"
          }
        },
        "required": [
          "models",
          "default",
          "default_provider"
        ],
        "title": "ChatModelsResponse",
        "type": "object"
      },
      "ChatStreamSchema": {
        "description": "Schema-only container that makes the stream's custom types appear in\nthe OpenAPI document. Never actually served.",
        "properties": {
          "data_part": {
            "$ref": "#/components/schemas/OpenPaperDataPart"
          },
          "message_metadata": {
            "$ref": "#/components/schemas/ChatMessageMetadata"
          }
        },
        "required": [
          "data_part",
          "message_metadata"
        ],
        "title": "ChatStreamSchema",
        "type": "object"
      },
...
      "LLMProvider": {
        "enum": [
          "gemini",
          "openai",
          "codex_proxy",
          "anthropic"
        ],
        "title": "LLMProvider",
        "type": "string"
      },
...
      "OpenPaperDataPart": {
        "description": "Every custom `data-*` part/chunk OpenPaper puts on the chat stream.",
        "discriminator": {
          "mapping": {
            "data-citations": "#/components/schemas/CitationsDataPart",
            "data-retry-status": "#/components/schemas/RetryStatusDataPart"
          },
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/CitationsDataPart"
          },
          {
            "$ref": "#/components/schemas/RetryStatusDataPart"
          }
        ],
        "title": "OpenPaperDataPart"
      },
...
      "PaperChatRequest": {
        "description": "The whole POST body: the AI SDK submit payload plus OpenPaper's fields.\n\n`messages` carries only the NEW user message; server-side history is\nground truth.",
        "properties": {
          "context_mode": {
            "anyOf": [
              {
                "enum": [
                  "adaptive",
                  "comprehensive",
                  "full"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Mode"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "llm_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Provider"
          },
          "messageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messageid"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/UIMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "paper_id": {
            "title": "Paper Id",
            "type": "string"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "xhigh",
                  "max"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "trigger": {
            "default": "submit-message",
            "enum": [
              "submit-message",
              "regenerate-message"
            ],
            "title": "Trigger",
            "type": "string"
          },
          "user_references": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "User References"
          }
        },
        "required": [
          "paper_id",
          "conversation_id",
          "id",
          "messages"
        ],
        "title": "PaperChatRequest",
        "type": "object"
      },
...
      "QuickQuestionCodeBody": {
        "description": "Inline code question. Ephemeral: no conversation, nothing persisted.",
        "properties": {
          "end_line": {
            "title": "End Line",
            "type": "integer"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "llm_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "paper_id": {
            "title": "Paper Id",
            "type": "string"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "xhigh",
                  "max"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          },
          "start_line": {
            "title": "Start Line",
            "type": "integer"
          }
        },
        "required": [
          "paper_id",
          "question",
          "file_path",
          "start_line",
          "end_line"
        ],
        "title": "QuickQuestionCodeBody",
        "type": "object"
      },
      "ReasoningUIPart": {
        "additionalProperties": false,
        "description": "A reasoning part of a message.",
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "providerMetadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providermetadata"
          },
          "state": {
            "anyOf": [
              {
                "enum": [
                  "streaming",
                  "done"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "const": "reasoning",
            "default": "reasoning",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ReasoningUIPart",
        "type": "object"
      },
...
      "TextUIPart": {
        "additionalProperties": false,
        "description": "A text part of a message.",
        "properties": {
          "providerMetadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providermetadata"
          },
          "state": {
            "anyOf": [
              {
                "enum": [
                  "streaming",
                  "done"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "const": "text",
            "default": "text",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TextUIPart",
        "type": "object"
      },
...
      "ToolInputAvailablePart": {
        "additionalProperties": false,
        "description": "Tool part in input-available state.",
        "properties": {
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolApprovalRequested"
              },
              {
                "$ref": "#/components/schemas/ToolApprovalResponded"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval"
          },
          "callProviderMetadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callprovidermetadata"
          },
          "input": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "providerExecuted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providerexecuted"
          },
          "state": {
            "const": "input-available",
            "default": "input-available",
            "title": "State",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "toolCallId": {
            "title": "Toolcallid",
            "type": "string"
          },
          "type": {
            "pattern": "^tool-",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "toolCallId"
        ],
        "title": "ToolInputAvailablePart",
        "type": "object"
      },
      "ToolInputStreamingPart": {
        "additionalProperties": false,
        "description": "Tool part in input-streaming state.",
        "properties": {
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolApprovalRequested"
              },
              {
                "$ref": "#/components/schemas/ToolApprovalResponded"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval"
          },
          "callProviderMetadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callprovidermetadata"
          },
          "input": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "providerExecuted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providerexecuted"
          },
          "state": {
            "const": "input-streaming",
            "default": "input-streaming",
            "title": "State",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "toolCallId": {
            "title": "Toolcallid",
            "type": "string"
          },
          "type": {
            "pattern": "^tool-",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "toolCallId"
        ],
        "title": "ToolInputStreamingPart",
        "type": "object"
      },
...
      "UIMessage": {
        "additionalProperties": false,
        "description": "A message as displayed in the UI by Vercel AI Elements.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "parts": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TextUIPart"
                },
                {
                  "$ref": "#/components/schemas/ReasoningUIPart"
                },
                {
                  "$ref": "#/components/schemas/ToolInputStreamingPart"
                },
                {
                  "$ref": "#/components/schemas/ToolInputAvailablePart"
                },
                {
                  "$ref": "#/components/schemas/ToolOutputAvailablePart"
                },
                {
                  "$ref": "#/components/schemas/ToolOutputErrorPart"
                },
                {
                  "$ref": "#/components/schemas/ToolApprovalRequestedPart"
                },
                {
                  "$ref": "#/components/schemas/ToolApprovalRespondedPart"
                },
                {
                  "$ref": "#/components/schemas/ToolOutputDeniedPart"
                },
                {
                  "$ref": "#/components/schemas/DynamicToolInputStreamingPart"
                },
                {
                  "$ref": "#/components/schemas/DynamicToolInputAvailablePart"
                },
                {
                  "$ref": "#/components/schemas/DynamicToolOutputAvailablePart"
                },
                {
                  "$ref": "#/components/schemas/DynamicToolOutputErrorPart"
                },
                {
                  "$ref": "#/components/schemas/DynamicToolApprovalRequestedPart"
                },
                {
                  "$ref": "#/components/schemas/DynamicToolApprovalRespondedPart"
                },
                {
                  "$ref": "#/components/schemas/DynamicToolOutputDeniedPart"
                },
                {
                  "$ref": "#/components/schemas/SourceUrlUIPart"
                },
                {
                  "$ref": "#/components/schemas/SourceDocumentUIPart"
                },
                {
                  "$ref": "#/components/schemas/FileUIPart"
                },
                {
                  "$ref": "#/components/schemas/DataUIPart"
                },
                {
                  "$ref": "#/components/schemas/StepStartUIPart"
                }
              ]
            },
            "title": "Parts",
            "type": "array"
          },
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "parts"
        ],
        "title": "UIMessage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A web application for uploading and annotating papers.",
    "title": "Open Paper",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/message/chat/paper": {
      "post": {
        "operationId": "chat_paper_api_message_chat_paper_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Vercel AI SDK UIMessage stream (SSE). OpenPaper's custom parts are typed by GET /api/message/stream-parts."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Server error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Chat Paper"
      }
    },
    "/api/message/models": {
      "get": {
        "description": "User-selectable chat models with capabilities, for the picker.",
        "operationId": "get_available_models_api_message_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Get Available Models"
      }
    },
    "/api/message/quick-question/code": {
      "post": {
        "description": "Stream a one-shot answer about a selected range of a repo file.\n\nSame wire protocol as /chat/paper (the client uses the identical ai-sdk\nreader), with read-only repo lookup tools but no citations and nothing\nwritten to conversations or messages.",
        "operationId": "quick_question_code_api_message_quick_question_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickQuestionCodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Vercel AI SDK UIMessage stream (SSE). OpenPaper's custom parts are typed by GET /api/message/stream-parts."
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Server error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Quick Question Code"
      }
    },
    "/api/message/stream-parts": {
      "get": {
        "description": "Schema-only: types OpenPaper's custom parts on the chat and\nquick-question UIMessage streams (`data-citations`, the transient\n`data-retry-status`, and assistant message metadata). Nothing to fetch;\nalways 404.",
        "operationId": "stream_parts_schema_api_message_stream_parts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatStreamSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Client error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Stream Parts Schema"
      }
    },
...
  }
}